Winston log integration and log rotation features
This commit is a consequence of another commit which involves the
forever-monito repo.  The spawner of the monitor process no longer has
to provide a file descriptor since the logFile, outFile and errFile are
all managed by the forever-monitor process that uses Winston and its
logrotate options. A logs-test is provided to check the log rotation
features and also an update to the readme and help that explain the new
options and their usage.